réorganisation des classes pour que ce soit plus parlant et/ou plus logique changement de la gestion de la puissance des gems en fonction de leur niveau
