+++ src/main/java/com/hyratrion/rpgnloots/event/ModEventToolTips.java
@@ -3,22 +3,17 @@ package com.hyratrion.rpgnloots.event;
 import com.google.common.collect.Multimap;
 import com.hyratrion.rpgnloots.RPGNLOOT;
 import com.hyratrion.rpgnloots.event.loot.CustomAttributes;
+import com.hyratrion.rpgnloots.item.Gems;
 import com.hyratrion.rpgnloots.item.ModItems;
-import com.hyratrion.rpgnloots.screen.ModMenuTypes;
-import com.hyratrion.rpgnloots.screen.SocketingTableScreen;
 import com.hyratrion.rpgnloots.util.ModTags;
 import com.hyratrion.rpgnloots.util.StaticClass;
 import net.minecraft.ChatFormatting;
-import net.minecraft.client.gui.screens.MenuScreens;
 import net.minecraft.client.gui.screens.Screen;
 import net.minecraft.core.Registry;
 import net.minecraft.locale.Language;
 import net.minecraft.nbt.CompoundTag;
 import net.minecraft.nbt.ListTag;
 import net.minecraft.network.chat.*;
-import net.minecraft.tags.TagKey;
-import net.minecraft.world.damagesource.DamageSource;
-import net.minecraft.world.entity.Entity;
 import net.minecraft.world.entity.EquipmentSlot;
 import net.minecraft.world.entity.MobType;
 import net.minecraft.world.entity.ai.attributes.Attribute;
@@ -27,17 +22,9 @@ import net.minecraft.world.entity.ai.attributes.Attributes;
 import net.minecraft.world.entity.player.Player;
 import net.minecraft.world.item.*;
 import net.minecraft.world.item.enchantment.EnchantmentHelper;
-import net.minecraftforge.event.RegisterCommandsEvent;
-import net.minecraftforge.event.entity.living.LivingDamageEvent;
-import net.minecraftforge.event.entity.living.LivingHurtEvent;
-import net.minecraftforge.event.entity.player.CriticalHitEvent;
 import net.minecraftforge.event.entity.player.ItemTooltipEvent;
-import net.minecraftforge.event.entity.player.PlayerEvent;
-import net.minecraftforge.eventbus.api.Event;
 import net.minecraftforge.eventbus.api.SubscribeEvent;
 import net.minecraftforge.fml.common.Mod;
-import net.minecraftforge.fml.event.lifecycle.FMLClientSetupEvent;
-import net.minecraftforge.server.command.ConfigCommand;
 
 import java.lang.reflect.Method;
 import java.util.*;
@@ -45,300 +32,8 @@ import java.util.*;
 import static net.minecraft.world.item.ItemStack.ATTRIBUTE_MODIFIER_FORMAT;
 
 @Mod.EventBusSubscriber(modid = RPGNLOOT.MOD_ID)
-public class ModEvents
+public class ModEventToolTips
 {
-    private static Random rand = new Random();
-
-    @SubscribeEvent
-    public static void onCommandsRegister(RegisterCommandsEvent event)
-    {
-        ConfigCommand.register(event.getDispatcher());
-    }
-
-    @SubscribeEvent
-    public static void onPlayerCloneEvent(PlayerEvent.Clone event)
-    {
-        if(!event.getOriginal().getLevel().isClientSide())
-        {
-            event.getPlayer().getPersistentData().putIntArray(RPGNLOOT.MOD_ID + "homepos",
-                    event.getOriginal().getPersistentData().getIntArray(RPGNLOOT.MOD_ID + "homepos"));
-        }
-    }
-
-    //event de critique
-    @SubscribeEvent
-    public static void onCriticalHitEvent(CriticalHitEvent event)
-    {
-        if(!event.getEntity().level.isClientSide())
-        {
-            Player player = event.getPlayer();
-            //récupération de l'item dans main du joueur
-            ItemStack itemStack = player.getMainHandItem();
-
-            //récupération des attributeModifiers de l'item
-            Multimap<Attribute, AttributeModifier> attributeModifiers = itemStack.getAttributeModifiers(EquipmentSlot.MAINHAND);
-
-            boolean gemLevelIncrease = attributeModifiers.containsKey(CustomAttributes.GEM_LVL_INCREASE.get());
-
-            //System.out.println("-- RPG&Loots -- player name => " + player.getDisplayName().getString());
-
-            //chance critique
-
-            float criticalChanceAttributeArmor = getValueAttributeOfArmor(CustomAttributes.CRITICAL_CHANCE.get(), player);
-            //System.out.println("- RPG&Loots - crit chance attribute armor => " + criticalChanceAttributeArmor);
-
-            float criticalChanceGemArmor = getValueGemOfArmor(ModTags.Items.GEM_TYPE_CRITICAL_CHANCE, player);
-            //System.out.println("- RPG&Loots - crit chance gem armor => " + criticalChanceGemArmor);
-
-            float criticalChance = criticalChanceAttributeArmor + criticalChanceGemArmor;
-
-            //récupération des chances de critiques des attributs de l'arme
-            if(attributeModifiers.containsKey(CustomAttributes.CRITICAL_CHANCE.get()))
-            {
-                float criticalChanceAttributeWeapon = StaticClass.GetValueFromAttributeModifierMap(attributeModifiers, CustomAttributes.CRITICAL_CHANCE.get());
-                //System.out.println("- RPG&Loots - crit chance attribute weapon => " + criticalChanceAttributeWeapon);
-                criticalChance += criticalChanceAttributeWeapon;
-            }
-
-
-            //récupération des chances de critiques des gems de l'arme
-            String[] allGemsEquiped = null;
-            if(ModTags.HasGemSlot(itemStack))
-            {
-                allGemsEquiped = ModTags.GetGemTags(itemStack);
-
-                if(ModTags.HasGemOfType(allGemsEquiped, ModTags.Items.GEM_TYPE_CRITICAL_CHANCE))
-                {
-                    float criticalChanceGemWeapon = ModTags.GetGemTotalValueOfType(itemStack, ModTags.Items.GEM_TYPE_CRITICAL_CHANCE, gemLevelIncrease);
-                    //System.out.println("- RPG&Loots - crit chance gem weapon => " + criticalChanceGemWeapon);
-                    criticalChance += criticalChanceGemWeapon;
-                }
-            }
-
-            //degats critique
-
-            //récupération de la valeur de chance d'appliquer un critique
-            float criticalChanceRNG = rand.nextFloat(100);
-
-            float criticalDamage = event.getOldDamageModifier() - 1;
-
-            //on check notre chance de faire un critique
-            if(criticalChanceRNG < criticalChance || criticalChance >= 100 || criticalDamage > 0)
-            {
-                int multiplicator = 1;
-                if(criticalChance >= 100)
-                {
-                    String critChanceStr = String.valueOf(criticalChance);
-                    int posComma = critChanceStr.indexOf(".");//par ce que 94 % sure de micro
-                    posComma = posComma != -1 ? posComma - 2 : 1;
-                    multiplicator += Integer.valueOf(critChanceStr.substring(0, posComma)); // ajoute +1 au multiplicateur critique par centaine
-                    //multiplicator += (Integer.valueOf(critChanceStr.substring(0, posComma)) / 2); // ajoute +0.5 au multiplicateur critique par centaine
-                    //System.out.println("- RPG&Loots - before multiplicator => " + multiplicator);
-
-                    criticalChanceRNG = rand.nextFloat(100);
-                    if(criticalChanceRNG < criticalChance - 100 * (multiplicator -1))
-                    {
-                        multiplicator += 1; // ajoute +1 au multiplicateur critique
-                        //multiplicator += 0.5f; // ajoute +0.5 au multiplicateur critique
-                    }
-                }
-                //System.out.println("- RPG&Loots - after multiplicator => " + multiplicator);
-
-
-                float criticalDamageAttributeArmor = getValueAttributeOfArmor(CustomAttributes.CRITICAL_DAMAGE.get(), player) / 100;
-                //System.out.println("- RPG&Loots - crit damage attribute armor => " + criticalDamageAttributeArmor);
-                criticalDamage += criticalDamageAttributeArmor;
-
-                float criticalDamageGemArmor = getValueGemOfArmor(ModTags.Items.GEM_TYPE_CRITICAL_DAMAGE, player) / 100;
-                //System.out.println("- RPG&Loots - crit damage gem armor => " + criticalDamageGemArmor);
-                criticalDamage += criticalDamageGemArmor;
-
-
-
-                //récupération des degats de critiques des attributes de l'arme
-                if(attributeModifiers.containsKey(CustomAttributes.CRITICAL_DAMAGE.get()))
-                {
-                    //on ajoute plus de dégâts critique
-                    float criticalDamgeWeapon = StaticClass.GetValueFromAttributeModifierMap(attributeModifiers, CustomAttributes.CRITICAL_DAMAGE.get()) / 100;
-                    //System.out.println("- RPG&Loots - crit damage attribute weapon => " + criticalDamgeWeapon);
-                    criticalDamage += criticalDamgeWeapon;
-                }
-
-
-                //récupération des degats de critiques des gems de l'arme
-                if(ModTags.HasGemSlot(itemStack))
-                {
-                    if(ModTags.HasGemOfType(allGemsEquiped, ModTags.Items.GEM_TYPE_CRITICAL_DAMAGE))
-                    {
-                        float criticalDamageGem = ModTags.GetGemTotalValueOfType(itemStack, ModTags.Items.GEM_TYPE_CRITICAL_DAMAGE, gemLevelIncrease) / 100;
-                        //System.out.println("- RPG&Loots - crit chance gem weapon => " + criticalDamageGem);
-                        criticalDamage += criticalDamageGem;
-                    }
-                }
-
-
-                criticalDamage = criticalDamage * multiplicator + 1;
-
-                //on change la valeur des dégâts
-                event.setDamageModifier(criticalDamage);
-                event.setResult(Event.Result.ALLOW);
-
-                //System.out.println("- RPG&Loots - multiplicator crit final => " + criticalDamage);
-            }
-            /*else
-            {
-                //System.out.println("----- Makotache ----- AUCUN critique");
-            }*/
-        }
-
-    }
-
-    //Event quand une enitité prend des dégats
-    @SubscribeEvent
-    public static void onLivingHurtEvent(LivingHurtEvent event)
-    {
-
-
-        if (!event.getEntity().level.isClientSide())
-        {
-            //si l'entité qui SUBIT des dégâts est un joueur
-            if (event.getEntity() instanceof Player player)
-            {
-                //on récupère tous les equipement dans les slots d'armure du joueur
-                Iterable<ItemStack> itemStacks = player.getArmorSlots();
-
-
-                boolean dodgeDone = false;
-
-                //esquive
-                float amountDodge = getValueAttributeOfArmor(CustomAttributes.DODGE.get(), player);
-
-                //si on a une chance de faire une equive
-                if (amountDodge > 0)
-                {
-                    //System.out.println("amout dodge -->" + amountDodge);
-                    //on test notre chance d'esquive
-                    float chanceDodge = rand.nextInt(100);
-                    if (chanceDodge < amountDodge)
-                    {
-                        dodgeDone = true;
-                        //si on esquive alors on ne prend pas les dégats
-                        event.setCanceled(true);
-                    }
-                }
-
-                //si on a pas fais d'esquive
-                if(!dodgeDone)
-                {
-                    float amountReflectedDamagePercent = 0;
-                    float amountReflectedDamageRaw = 0;
-
-                    //renvoie de degats
-                    for (ItemStack itemStack : itemStacks)
-                    {
-                        //si l'equipement d'amure est bien une armure et pas un slot vide
-                        if (itemStack.getItem() instanceof ArmorItem armorItem)
-                        {
-                            //System.out.println("Get equipement slot --> " + armorItem.getSlot());
-                            //récupération des attributeModifiers de l'item en question
-                            Multimap<Attribute, AttributeModifier> attributeModifiers = itemStack.getAttributeModifiers(armorItem.getSlot());
-
-                            //POURCENTAGE
-
-                            //si l'armure possède "CustomAttributes.REFLECT_DAMAGE_PERCENT"
-                            if (attributeModifiers.containsKey(CustomAttributes.REFLECT_DAMAGE_PERCENT.get()))
-                            {
-                                //récupérationd de la valeurs de renvoie de dégâts
-                                amountReflectedDamagePercent += StaticClass.GetValueFromAttributeModifierMap(attributeModifiers, CustomAttributes.REFLECT_DAMAGE_PERCENT.get()) / 100;
-                            }
-
-                            //BRUTE
-                            //si l'armure possède "CustomAttributes.REFLECT_DAMAGE_RAW"
-                            if (attributeModifiers.containsKey(CustomAttributes.REFLECT_DAMAGE_RAW.get()))
-                            {
-                                amountReflectedDamageRaw += StaticClass.GetValueFromAttributeModifierMap(attributeModifiers, CustomAttributes.REFLECT_DAMAGE_RAW.get());
-                            }
-                        }
-                    }
-
-                    //réucpération de l'entitié qui INFLIGE les dégats
-                    Entity entity = event.getSource().getDirectEntity();
-
-
-                    //si on a un renvoie de dégâts
-                    if(entity != null && (amountReflectedDamagePercent > 0 || amountReflectedDamageRaw > 0))
-                    {
-                        //calcule du montant de dégâs a infligé en pourcentage
-                        float damgeToReflect = amountReflectedDamagePercent * event.getAmount();
-
-                        //on ajotue le montant brute de dégâts
-                        damgeToReflect += amountReflectedDamageRaw;
-
-                        //on inflige les dégâts
-                        entity.hurt(DamageSource.playerAttack(player), damgeToReflect);
-                    }
-                }
-            }
-        }
-    }
-
-    //Event juste avnt entité prend des degats
-    //après le calcule le nombre de degats en fonction des resistance
-    @SubscribeEvent
-    public static void onLivingDamageEvent(LivingDamageEvent event)
-    {
-        if (!event.getEntity().level.isClientSide())
-        {
-            if (event.getSource().getDirectEntity() instanceof Player player)
-            {
-                //vol de vie POURCENTAGE
-                ItemStack itemStack = player.getMainHandItem();
-
-                Multimap<Attribute, AttributeModifier> attributeModifiers = itemStack.getAttributeModifiers(EquipmentSlot.MAINHAND);
-
-                //System.out.println("-- RPG&Loots -- bobo of " + event.getEntity().getType().getRegistryName().getPath() + " => " + event.getAmount());
-
-                //vol de vie
-                if(attributeModifiers.containsKey(CustomAttributes.LIFE_LEECH_PERCENT.get()))
-                {
-                    float amountLifeLeechPercent = StaticClass.GetValueFromAttributeModifierMap(attributeModifiers, CustomAttributes.LIFE_LEECH_PERCENT.get()) / 100;
-
-                    //System.out.println("----- Makotache ----- pourcentage de vole de vie => " + amountLifeLeechPercent);
-
-                    amountLifeLeechPercent = amountLifeLeechPercent * event.getAmount();
-
-                    player.heal(amountLifeLeechPercent);
-
-                    //System.out.println("----- Makotache ----- vie regen => " + amountLifeLeechPercent);
-                }
-
-
-                //vol de vie BRUTE
-
-                //System.out.println("----- Makotache ----- bobo du mob => " +event.getAmount());
-
-                //Component displayname = itemStack.getDisplayName();
-                //System.out.println("----- Makotache ----- displayname => " + displayname);
-                //String DescriptionId = itemStack.getDescriptionId();
-                //System.out.println("----- Makotache ----- DescriptionId => " + DescriptionId);
-
-                if(attributeModifiers.containsKey(CustomAttributes.LIFE_LEECH_RAW.get()))
-                {
-                    float amountLifeLeechraw = StaticClass.GetValueFromAttributeModifierMap(attributeModifiers, CustomAttributes.LIFE_LEECH_RAW.get());
-
-                    //System.out.println("----- Makotache ----- pourcentage de vole de vie => " + amountLifeLeechraw);
-
-                    player.heal(amountLifeLeechraw);
-
-                    //System.out.println("----- Makotache ----- vie regen => " + amountLifeLeechraw);
-                }
-            }
-        }
-    }
-
-
-
-
     @SubscribeEvent
     public static void onItemTooltipEvent(ItemTooltipEvent event)
     {
@@ -360,7 +55,7 @@ public class ModEvents
 
                 ChatFormatting color = null;
 
-                switch (ModTags.GetTier(itemStack))
+                switch (Gems.GetTier(itemStack))
                 {
                     case 1 :
                         color = ChatFormatting.GREEN;
@@ -447,7 +142,7 @@ public class ModEvents
                     }
                     else
                     {
-                        equipmentSlots = ModTags.GetEquipmentSlotOf(itemStack);
+                        equipmentSlots = StaticClass.GetEquipmentSlotOf(itemStack);
                     }
                     //EquipmentSlot[] equipmentSlots = ModTags.GetEquipmentSlotOf(itemStack);
                     for(int i =0; i < equipmentSlots.length; i++)
@@ -501,7 +196,7 @@ public class ModEvents
                                 else if(attribute.equals(CustomAttributes.CRITICAL_DAMAGE.get()))
                                 {
                                     showPercent = true;
-                                    chatFormatting = ChatFormatting.GOLD;//ModColor.RED_ORANGE;
+                                    chatFormatting = ChatFormatting.GOLD;
                                 }
                                 else if(attribute.equals(Attributes.MAX_HEALTH))
                                 {
@@ -529,7 +224,7 @@ public class ModEvents
 
                                         chatFormatting = ChatFormatting.GRAY;
 
-                                        Item[] allGemsEquiped = ModTags.GetGems(itemStack);
+                                        Item[] allGemsEquiped = Gems.GetGems(itemStack);
 
                                         //on doit appuyer sur shift pou plus de détail
                                         if(Screen.hasShiftDown())
@@ -545,9 +240,9 @@ public class ModEvents
                                                 }
                                                 else
                                                 {
-                                                    showPercent = ModTags.GemIsType(item, ModTags.Items.GEM_PERCENT);
+                                                    showPercent = Gems.GemIsType(item, ModTags.Items.GEM_PERCENT);
 
-                                                    componentToolTip = createComponentToolTipGem(gemText + ub, ModTags.GetGemValue(item), item, showPercent, chatFormatting, gemLevelIncrease);
+                                                    componentToolTip = createComponentToolTipGem(gemText + ub, Gems.GetGemValue(item), item, showPercent, chatFormatting, gemLevelIncrease);
                                                 }
                                                 unSortedAttributeModifierMaps.put(componentToolTip.getKey(), componentToolTip.getValue());
                                             }
@@ -721,7 +416,8 @@ public class ModEvents
 
         if(gemLevelIncrease)
         {
-            text += "(+" + ATTRIBUTE_MODIFIER_FORMAT.format(ModTags.GetGemValue(ModTags.GetGemType(item), 1)) + ") ";
+            float valueGemLevelIncrease = Gems.GetGemValue(item, Gems.GetGemLevel(item) + 1) - Gems.GetGemValue(item, Gems.GetGemLevel(item));
+            text += "(+" + ATTRIBUTE_MODIFIER_FORMAT.format(valueGemLevelIncrease) + ") ";
         }
 
         return createComponentToolTipGem(pos, text, item.getDescriptionId(), chatFormatting);
@@ -733,93 +429,6 @@ public class ModEvents
 
     private static int getHideFlags(ItemStack itemStack) {
         return itemStack.hasTag() && itemStack.getTag().contains("HideFlags", 99) ? itemStack.getTag().getInt("HideFlags") : itemStack.getItem().getDefaultTooltipHideFlags(itemStack);
-    }
-
-
-    private static float getValueAttributeOfArmor(Attribute attribute, Player player)
-    {
-        float result = 0;
-
-        ItemStack offHand = player.getOffhandItem();
-
-        if(!offHand.isEmpty())
-        {
-            Multimap<Attribute, AttributeModifier> attributeModifiers = offHand.getAttributeModifiers(EquipmentSlot.OFFHAND);
-
-            //si attributeModifiers contient "CustomAttributes.DODGE.get())"
-            if (attributeModifiers.containsKey(attribute))
-            {
-                //on réucpère le montant d'esquive
-                result += StaticClass.GetValueFromAttributeModifierMap(attributeModifiers, attribute);
-            }
-        }
-
-        Iterable<ItemStack> itemStacks = player.getArmorSlots();
-
-        for (ItemStack itemStack : itemStacks)
-        {
-            //si l'equipement d'amure est bien une armure et pas un slot vide
-            if (itemStack.getItem() instanceof ArmorItem armorItem)
-            {
-                //récupération des attributeModifiers de l'item en question
-                Multimap<Attribute, AttributeModifier> attributeModifiers = itemStack.getAttributeModifiers(armorItem.getSlot());
-
-                //si attributeModifiers contient "CustomAttributes.DODGE.get())"
-                if (attributeModifiers.containsKey(attribute))
-                {
-                    //on réucpère le montant d'esquive
-                    result += StaticClass.GetValueFromAttributeModifierMap(attributeModifiers, attribute);
-                }
-            }
-        }
-
-        return result;
-    }
-
-    private static float getValueGemOfArmor(TagKey<Item> gemType, Player player)
-    {
-        float result = 0;
-
-        ItemStack offHand = player.getOffhandItem();
-
-
-        if(!offHand.isEmpty() && ModTags.HasGemSlot(offHand))
-        {
-            String[] allGemsEquiped = ModTags.GetGemTags(offHand);
-
-
-            if(ModTags.HasGemOfType(allGemsEquiped, gemType))
-            {
-                Multimap<Attribute, AttributeModifier> attributeModifiers = offHand.getAttributeModifiers(EquipmentSlot.OFFHAND);
-                boolean gemLevelIncrease = attributeModifiers.containsKey(CustomAttributes.GEM_LVL_INCREASE.get());
-
-                result += ModTags.GetGemTotalValueOfType(offHand, gemType, gemLevelIncrease);
-            }
-        }
-
-        Iterable<ItemStack> itemStacks = player.getArmorSlots();
-
-        for (ItemStack itemStack : itemStacks)
-        {
-            //si l'equipement d'amure est bien une armure et pas un slot vide
-            if (itemStack.getItem() instanceof ArmorItem armorItem)
-            {
-                if(ModTags.HasGemSlot(itemStack))
-                {
-                    String[] allGemsEquiped = ModTags.GetGemTags(itemStack);
-
-                    if(ModTags.HasGemOfType(allGemsEquiped, gemType))
-                    {
-                        Multimap<Attribute, AttributeModifier> attributeModifiers = itemStack.getAttributeModifiers(armorItem.getSlot());
-                        boolean gemLevelIncrease = attributeModifiers.containsKey(CustomAttributes.GEM_LVL_INCREASE.get());
-
-                        result += ModTags.GetGemTotalValueOfType(itemStack, gemType, gemLevelIncrease);
-                    }
-                }
-            }
-        }
-
-        return result;
     }
 
 }
