+++ src/main/java/com/hyratrion/rpgnloots/item/ModCreativeTab.java
@@ -3,7 +3,8 @@ package com.hyratrion.rpgnloots.item;
 import net.minecraft.world.item.CreativeModeTab;
 import net.minecraft.world.item.ItemStack;
 
-public class ModCreativeModeTab {
+public class ModCreativeTab
+{
     public static final CreativeModeTab RPGNLOOTS_TAB = new CreativeModeTab("rpgnlootstab") {
         @Override
         public ItemStack makeIcon() {
