+++ src/main/java/com/hyratrion/rpgnloots/item/ModItems.java
@@ -3,7 +3,6 @@ package com.hyratrion.rpgnloots.item;
 import com.hyratrion.rpgnloots.RPGNLOOT;
 import com.hyratrion.rpgnloots.block.ModBlocks;
 import com.hyratrion.rpgnloots.util.ModTags;
-import net.minecraft.nbt.StringTag;
 import net.minecraft.network.chat.Component;
 import net.minecraft.network.chat.TextComponent;
 import net.minecraft.network.chat.TranslatableComponent;
@@ -29,13 +28,13 @@ public class ModItems
             DeferredRegister.create(ForgeRegistries.ITEMS, RPGNLOOT.MOD_ID);
 
     public static final RegistryObject<Item> PLIERS = ITEMS.register("pliers",
-            () -> new Item(new Item.Properties().tab(ModCreativeModeTab.RPGNLOOTS_TAB)));
+            () -> new Item(new Item.Properties().tab(ModCreativeTab.RPGNLOOTS_TAB)));
 
-    public static final RegistryObject<Item> SOCKETING_TABLE = block(ModBlocks.SOCKETING_TABLE, ModCreativeModeTab.RPGNLOOTS_TAB);
+    public static final RegistryObject<Item> SOCKETING_TABLE = block(ModBlocks.SOCKETING_TABLE, ModCreativeTab.RPGNLOOTS_TAB);
 
     //Gemmes lvl 0
     public static final RegistryObject<Item> GEM_CHANCE_LVL_0 = ITEMS.register("gem_chance_lvl_0",
-            () -> new Item(new Item.Properties().tab(ModCreativeModeTab.RPGNLOOTS_TAB))
+            () -> new Item(new Item.Properties().tab(ModCreativeTab.RPGNLOOTS_TAB))
             {
                 @Override
                 public void appendHoverText(ItemStack itemStack, @Nullable Level level, List<Component> components, TooltipFlag tooltipFlag)
@@ -45,7 +44,7 @@ public class ModItems
             });
 
     public static final RegistryObject<Item> GEM_CRITICAL_CHANCE_LVL_0 = ITEMS.register("gem_critical_chance_lvl_0",
-            () -> new Item(new Item.Properties().tab(ModCreativeModeTab.RPGNLOOTS_TAB))
+            () -> new Item(new Item.Properties().tab(ModCreativeTab.RPGNLOOTS_TAB))
             {
                 @Override
                 public void appendHoverText(ItemStack itemStack, @Nullable Level level, List<Component> components, TooltipFlag tooltipFlag)
@@ -55,7 +54,7 @@ public class ModItems
             });
 
     public static final RegistryObject<Item> GEM_CRITICAL_DAMAGE_LVL_0 = ITEMS.register("gem_critical_damage_lvl_0",
-            () -> new Item(new Item.Properties().tab(ModCreativeModeTab.RPGNLOOTS_TAB))
+            () -> new Item(new Item.Properties().tab(ModCreativeTab.RPGNLOOTS_TAB))
             {
                 @Override
                 public void appendHoverText(ItemStack itemStack, @Nullable Level level, List<Component> components, TooltipFlag tooltipFlag)
@@ -65,7 +64,7 @@ public class ModItems
             });
 
     public static final RegistryObject<Item> GEM_DAMAGE_LVL_0 = ITEMS.register("gem_damage_lvl_0",
-            () -> new Item(new Item.Properties().tab(ModCreativeModeTab.RPGNLOOTS_TAB))
+            () -> new Item(new Item.Properties().tab(ModCreativeTab.RPGNLOOTS_TAB))
             {
                 @Override
                 public void appendHoverText(ItemStack itemStack, @Nullable Level level, List<Component> components, TooltipFlag tooltipFlag)
@@ -75,7 +74,7 @@ public class ModItems
             });
 
     public static final RegistryObject<Item> GEM_MAX_LIFE_LVL_0 = ITEMS.register("gem_max_life_lvl_0",
-            () -> new Item(new Item.Properties().tab(ModCreativeModeTab.RPGNLOOTS_TAB))
+            () -> new Item(new Item.Properties().tab(ModCreativeTab.RPGNLOOTS_TAB))
             {
                 @Override
                 public void appendHoverText(ItemStack itemStack, @Nullable Level level, List<Component> components, TooltipFlag tooltipFlag)
@@ -85,7 +84,7 @@ public class ModItems
             });
 
     public static final RegistryObject<Item> GEM_REPARATION_LVL_0 = ITEMS.register("gem_reparation_lvl_0",
-            () -> new Item(new Item.Properties().tab(ModCreativeModeTab.RPGNLOOTS_TAB))
+            () -> new Item(new Item.Properties().tab(ModCreativeTab.RPGNLOOTS_TAB))
             {
                 @Override
                 public void appendHoverText(ItemStack itemStack, @Nullable Level level, List<Component> components, TooltipFlag tooltipFlag)
@@ -95,7 +94,7 @@ public class ModItems
             });
 
     public static final RegistryObject<Item> GEM_SHIELD_LVL_0 = ITEMS.register("gem_shield_lvl_0",
-            () -> new Item(new Item.Properties().tab(ModCreativeModeTab.RPGNLOOTS_TAB))
+            () -> new Item(new Item.Properties().tab(ModCreativeTab.RPGNLOOTS_TAB))
             {
                 @Override
                 public void appendHoverText(ItemStack itemStack, @Nullable Level level, List<Component> components, TooltipFlag tooltipFlag)
@@ -105,7 +104,7 @@ public class ModItems
             });
 
     public static final RegistryObject<Item> GEM_SPEED_LVL_0 = ITEMS.register("gem_speed_lvl_0",
-            () -> new Item(new Item.Properties().tab(ModCreativeModeTab.RPGNLOOTS_TAB))
+            () -> new Item(new Item.Properties().tab(ModCreativeTab.RPGNLOOTS_TAB))
             {
                 @Override
                 public void appendHoverText(ItemStack itemStack, @Nullable Level level, List<Component> components, TooltipFlag tooltipFlag)
@@ -203,14 +202,7 @@ public class ModItems
 
     public static final RegistryObject<Item> GEM_SPEED_LVL_5 = GemRegister("gem_speed_lvl_5");
 
-    public static Item[] GEM_LVL_0;
-    public static Item[] GEM_LVL_1;
-    public static Item[] GEM_LVL_2;
-    public static Item[] GEM_LVL_3;
-    public static Item[] GEM_LVL_4;
-    public static Item[] GEM_LVL_5;
 
-    public static Map<Item, Float> GEMS_VALUES = new HashMap<>();
     public static Map<String, Item> GEMS_REFERENCES = new HashMap<>();
 
     public static final String GEM_TYPE = "gem_type";
@@ -218,7 +210,7 @@ public class ModItems
     private static RegistryObject<Item> GemRegister(String name)
     {
         return ITEMS.register(name,
-            () -> new Item(new Item.Properties().tab(ModCreativeModeTab.RPGNLOOTS_TAB))
+            () -> new Item(new Item.Properties().tab(ModCreativeTab.RPGNLOOTS_TAB))
             {
                 @Override
                 public void appendHoverText(ItemStack itemStack, @Nullable Level level, List<Component> components, TooltipFlag tooltipFlag)
@@ -242,7 +234,7 @@ public class ModItems
     private static Component gemDescription(Item gem)
     {
         Component result;
-        TagKey<Item> itemType = ModTags.GetGemType(gem);
+        TagKey<Item> itemType = Gems.GetGemType(gem);
 
         if(itemType != null)
         {
@@ -280,9 +272,9 @@ public class ModItems
                 text_gem_type = "tooltip.rpgnloots.gem_of_speed";
             }
 
-            float value = ModTags.GetGemValue(gem);
+            float value = Gems.GetGemValue(gem);
 
-            boolean showPercent = ModTags.GemIsType(gem, ModTags.Items.GEM_PERCENT);
+            boolean showPercent = Gems.GemIsType(gem, ModTags.Items.GEM_PERCENT);
 
             String percentSpace = showPercent ? "% " : " ";
 
@@ -298,54 +290,6 @@ public class ModItems
 
     public static void initGemsValuesAndReferences()
     {
-        //GEMS_VALUES
-
-        GEMS_VALUES.put(GEM_CHANCE_LVL_1.get(), 1f);
-        GEMS_VALUES.put(GEM_CRITICAL_CHANCE_LVL_1.get(), 15f);
-        GEMS_VALUES.put(GEM_CRITICAL_DAMAGE_LVL_1.get(), 15f);
-        GEMS_VALUES.put(GEM_DAMAGE_LVL_1.get(), 15f);
-        GEMS_VALUES.put(GEM_MAX_LIFE_LVL_1.get(), 1f);
-        GEMS_VALUES.put(GEM_REPARATION_LVL_1.get(), 1f);
-        GEMS_VALUES.put(GEM_SHIELD_LVL_1.get(), 1f);
-        GEMS_VALUES.put(GEM_SPEED_LVL_1.get(), 15f);
-
-        GEMS_VALUES.put(GEM_CHANCE_LVL_2.get(), 2f);
-        GEMS_VALUES.put(GEM_CRITICAL_CHANCE_LVL_2.get(), 30f);
-        GEMS_VALUES.put(GEM_CRITICAL_DAMAGE_LVL_2.get(), 30f);
-        GEMS_VALUES.put(GEM_DAMAGE_LVL_2.get(), 30f);
-        GEMS_VALUES.put(GEM_MAX_LIFE_LVL_2.get(), 2f);
-        GEMS_VALUES.put(GEM_REPARATION_LVL_2.get(), 2f);
-        GEMS_VALUES.put(GEM_SHIELD_LVL_2.get(), 2f);
-        GEMS_VALUES.put(GEM_SPEED_LVL_2.get(), 30f);
-
-        GEMS_VALUES.put(GEM_CHANCE_LVL_3.get(), 3f);
-        GEMS_VALUES.put(GEM_CRITICAL_CHANCE_LVL_3.get(), 45f);
-        GEMS_VALUES.put(GEM_CRITICAL_DAMAGE_LVL_3.get(), 45f);
-        GEMS_VALUES.put(GEM_DAMAGE_LVL_3.get(), 45f);
-        GEMS_VALUES.put(GEM_MAX_LIFE_LVL_3.get(), 3f);
-        GEMS_VALUES.put(GEM_REPARATION_LVL_3.get(), 3f);
-        GEMS_VALUES.put(GEM_SHIELD_LVL_3.get(), 3f);
-        GEMS_VALUES.put(GEM_SPEED_LVL_3.get(), 45f);
-
-        GEMS_VALUES.put(GEM_CHANCE_LVL_4.get(), 4f);
-        GEMS_VALUES.put(GEM_CRITICAL_CHANCE_LVL_4.get(), 70f);
-        GEMS_VALUES.put(GEM_CRITICAL_DAMAGE_LVL_4.get(), 70f);
-        GEMS_VALUES.put(GEM_DAMAGE_LVL_4.get(), 70f);
-        GEMS_VALUES.put(GEM_MAX_LIFE_LVL_4.get(), 4f);
-        GEMS_VALUES.put(GEM_REPARATION_LVL_4 .get(), 4f);
-        GEMS_VALUES.put(GEM_SHIELD_LVL_4.get(), 4f);
-        GEMS_VALUES.put(GEM_SPEED_LVL_4.get(), 70f);
-
-        GEMS_VALUES.put(GEM_CHANCE_LVL_5.get(), 5f);
-        GEMS_VALUES.put(GEM_CRITICAL_CHANCE_LVL_5.get(), 100f);
-        GEMS_VALUES.put(GEM_CRITICAL_DAMAGE_LVL_5.get(), 100f);
-        GEMS_VALUES.put(GEM_DAMAGE_LVL_5.get(), 100f);
-        GEMS_VALUES.put(GEM_MAX_LIFE_LVL_5.get(), 5f);
-        GEMS_VALUES.put(GEM_REPARATION_LVL_5.get(), 5f);
-        GEMS_VALUES.put(GEM_SHIELD_LVL_5.get(), 5f);
-        GEMS_VALUES.put(GEM_SPEED_LVL_5.get(), 100f);
-
-
         //GEMS_REFERENCES
 
         GEMS_REFERENCES.put(GEM_CHANCE_LVL_1.get().getRegistryName().getPath(), GEM_CHANCE_LVL_1.get());
@@ -392,79 +336,5 @@ public class ModItems
         GEMS_REFERENCES.put(GEM_REPARATION_LVL_5.get().getRegistryName().getPath(), GEM_REPARATION_LVL_5.get());
         GEMS_REFERENCES.put(GEM_SHIELD_LVL_5.get().getRegistryName().getPath(), GEM_SHIELD_LVL_5.get());
         GEMS_REFERENCES.put(GEM_SPEED_LVL_5.get().getRegistryName().getPath(), GEM_SPEED_LVL_5.get());
-    }
-
-    public static void createArrayGems()
-    {
-        GEM_LVL_0 = new Item[]
-        {
-            GEM_CHANCE_LVL_0.get(),
-            GEM_CRITICAL_CHANCE_LVL_0.get(),
-            GEM_CRITICAL_DAMAGE_LVL_0.get(),
-            GEM_DAMAGE_LVL_0.get(),
-            GEM_MAX_LIFE_LVL_0.get(),
-            GEM_REPARATION_LVL_0.get(),
-            GEM_SHIELD_LVL_0.get(),
-            GEM_SPEED_LVL_0.get()
-        };
-        GEM_LVL_1 = new Item[]
-        {
-            GEM_CHANCE_LVL_1.get(),
-            GEM_CRITICAL_CHANCE_LVL_1.get(),
-            GEM_CRITICAL_DAMAGE_LVL_1.get(),
-            GEM_DAMAGE_LVL_1.get(),
-            GEM_MAX_LIFE_LVL_1.get(),
-            GEM_REPARATION_LVL_1.get(),
-            GEM_SHIELD_LVL_1.get(),
-            GEM_SPEED_LVL_1.get()
-        };
-        
-        GEM_LVL_2 = new Item[]
-        {
-            GEM_CHANCE_LVL_2.get(),
-            GEM_CRITICAL_CHANCE_LVL_2.get(),
-            GEM_CRITICAL_DAMAGE_LVL_2.get(),
-            GEM_DAMAGE_LVL_2.get(),
-            GEM_MAX_LIFE_LVL_2.get(),
-            GEM_REPARATION_LVL_2.get(),
-            GEM_SHIELD_LVL_2.get(),
-            GEM_SPEED_LVL_2.get()
-        };
-
-        GEM_LVL_3 = new Item[]
-        {
-            GEM_CHANCE_LVL_3.get(),
-            GEM_CRITICAL_CHANCE_LVL_3.get(),
-            GEM_CRITICAL_DAMAGE_LVL_3.get(),
-            GEM_DAMAGE_LVL_3.get(),
-            GEM_MAX_LIFE_LVL_3.get(),
-            GEM_REPARATION_LVL_3.get(),
-            GEM_SHIELD_LVL_3.get(),
-            GEM_SPEED_LVL_3.get()
-        };
-
-        GEM_LVL_4 = new Item[]
-        {
-            GEM_CHANCE_LVL_4.get(),
-            GEM_CRITICAL_CHANCE_LVL_4.get(),
-            GEM_CRITICAL_DAMAGE_LVL_4.get(),
-            GEM_DAMAGE_LVL_4.get(),
-            GEM_MAX_LIFE_LVL_4.get(),
-            GEM_REPARATION_LVL_4.get(),
-            GEM_SHIELD_LVL_4.get(),
-            GEM_SPEED_LVL_4.get()
-        };
-
-        GEM_LVL_5 = new Item[]
-        {
-            GEM_CHANCE_LVL_5.get(),
-            GEM_CRITICAL_CHANCE_LVL_5.get(),
-            GEM_CRITICAL_DAMAGE_LVL_5.get(),
-            GEM_DAMAGE_LVL_5.get(),
-            GEM_MAX_LIFE_LVL_5.get(),
-            GEM_REPARATION_LVL_5.get(),
-            GEM_SHIELD_LVL_5.get(),
-            GEM_SPEED_LVL_5.get()
-        };
     }
 }
